Create seperate BacktestStrategy class
