+++ HFStrategy/__init__.py
@@ -1,6 +1,6 @@
 name = 'HFStrategy'
 
 from HFStrategy.Strategy.Strategy import Strategy
-from HFStrategy.utils.backtest import execOffline
+from HFStrategy.Strategy.BacktestStrategy import BacktestStrategy
 from HFStrategy.Strategy.PositionManager import PositionError
 from HFStrategy.Strategy.Position import Position
